add authentication checks to all credentials endpoints

+++ src/main/java/fi/digitraffic/tis/vaco/credentials/CredentialsController.java
@@ -8,6 +8,7 @@ import fi.digitraffic.tis.vaco.api.model.credentials.CreateCredentialsRequest;
 import fi.digitraffic.tis.vaco.api.model.credentials.UpdateCredentialsRequest;
 import fi.digitraffic.tis.vaco.credentials.model.Credentials;
 import fi.digitraffic.tis.vaco.credentials.model.CredentialsType;
+import fi.digitraffic.tis.vaco.me.MeService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -29,36 +30,46 @@ import java.util.Objects;
 public class CredentialsController {
 
     private final CredentialsService credentialsService;
+    private final MeService meService;
 
-    public CredentialsController(CredentialsService credentialsService) {
+    public CredentialsController(CredentialsService credentialsService, MeService meService) {
         this.credentialsService = Objects.requireNonNull(credentialsService);
+        this.meService = Objects.requireNonNull(meService);
     }
 
     @PostMapping(path = "")
     @JsonView(DataVisibility.AdminRestricted.class)
     @PreAuthorize("hasAuthority('vaco.admin')")
     public ResponseEntity<Resource<Credentials>> create(@Valid @RequestBody CreateCredentialsRequest credentials) {
-        return switch (credentials.type().fieldName()) {
-            case CredentialsType.Name.HTTP_BASIC -> credentialsService.createCredentials(credentials)
-                .map(Responses::created)
-                .orElseGet(() -> Responses.badRequest(
-                    String.format(
-                        "Failed to store credentials '%s'/'%s' for %s",
-                        credentials.name(),
-                        credentials.description(),
-                        credentials.owner()
-                    )));
-            default -> Responses.badRequest(String.format("Unsupported credentials type '%s'", credentials.type()));
-        };
+        if (meService.isAllowedToAccess(credentials.owner())) {
+            return switch (credentials.type().fieldName()) {
+                case CredentialsType.Name.HTTP_BASIC -> credentialsService.createCredentials(credentials)
+                    .map(Responses::created)
+                    .orElseGet(() -> Responses.badRequest(
+                        String.format(
+                            "Failed to store credentials '%s'/'%s' for %s",
+                            credentials.name(),
+                            credentials.description(),
+                            credentials.owner()
+                        )));
+                default -> Responses.badRequest(String.format("Unsupported credentials type '%s'", credentials.type()));
+            };
+        } else {
+            return Responses.unauthorized("Not allowed to access %s".formatted(credentials.owner()));
+        }
     }
 
     @GetMapping(path = "")
     @JsonView(DataVisibility.AdminRestricted.class)
     @PreAuthorize("hasAuthority('vaco.admin')")
     public ResponseEntity<Resource<List<Credentials>>> fetchAll(@RequestParam("businessId") String businessId) {
-        return credentialsService.findAllForBusinessId(businessId)
-            .map(Responses::ok)
-            .orElseGet(() -> Responses.badRequest(String.format("Could not fetch all credentials for '%s'", businessId)));
+        if (meService.isAllowedToAccess(businessId)) {
+            return credentialsService.findAllForBusinessId(businessId)
+                .map(Responses::ok)
+                .orElseGet(() -> Responses.badRequest(String.format("Could not fetch all credentials for '%s'", businessId)));
+        } else {
+            return Responses.unauthorized("Not allowed to access %s".formatted(businessId));
+        }
     }
 
     @GetMapping(path = "/{publicId}")
@@ -66,8 +77,16 @@ public class CredentialsController {
     @PreAuthorize("hasAuthority('vaco.admin')")
     public ResponseEntity<Resource<Credentials>> fetchCredentials(@PathVariable("publicId") String publicId) {
         return credentialsService.findByPublicId(publicId)
-            .map(Responses::ok)
-            .orElseGet(() -> Responses.badRequest(String.format("Could not fetch credentials with id '%s'", publicId)));
+            .map(c -> {
+                if (meService.isAllowedToAccess(c.owner())) {
+                    return credentialsService.findByPublicId(publicId)
+                        .map(Responses::ok)
+                        .orElseGet(() -> Responses.badRequest(String.format("Could not fetch credentials with id '%s'", publicId)));
+                } else {
+                    return Responses.<Credentials>unauthorized("Not allowed to access %s".formatted(publicId));
+                }
+            })
+            .orElseGet(() -> Responses.notFound("%s not found".formatted(publicId)));
     }
 
     @PutMapping(path = "/{publicId}")
@@ -75,19 +94,34 @@ public class CredentialsController {
     @PreAuthorize("hasAuthority('vaco.admin')")
     public ResponseEntity<Resource<Credentials>> updateCredentials(@PathVariable("publicId") String publicId,
                                                                    @Valid @RequestBody UpdateCredentialsRequest credentials) {
-        // TODO: this definitely needs access check
-        return credentialsService.updateCredentials(publicId, credentials)
-            .map(Responses::ok)
-            .orElseGet(() -> Responses.badRequest(String.format("Could not update credentials with id '%s'", publicId)));
+        return credentialsService.findByPublicId(publicId)
+            .map(c -> {
+                if (meService.isAllowedToAccess(c.owner())) {
+                    return credentialsService.updateCredentials(publicId, credentials)
+                        .map(Responses::ok)
+                        .orElseGet(() -> Responses.badRequest(String.format("Could not update credentials with id '%s'", publicId)));
+                } else {
+                    return Responses.<Credentials>unauthorized("Not allowed to access %s".formatted(publicId));
+                }
+            })
+            .orElseGet(() -> Responses.notFound("%s not found".formatted(publicId)));
     }
 
     @DeleteMapping(path = "/{publicId}")
     @JsonView(DataVisibility.AdminRestricted.class)
     @PreAuthorize("hasAuthority('vaco.admin')")
     public ResponseEntity<Resource<Boolean>> deleteCredentials(@PathVariable("publicId") String publicId) {
-        return credentialsService.deleteCredentials(publicId)
-            .map(Responses::ok)
-            .orElseGet(() -> Responses.badRequest(String.format("Could not delete credentials with id '%s'", publicId)));
+        return credentialsService.findByPublicId(publicId)
+            .map(c -> {
+                if (meService.isAllowedToAccess(c.owner())) {
+                    return credentialsService.deleteCredentials(publicId)
+                        .map(Responses::ok)
+                        .orElseGet(() -> Responses.badRequest(String.format("Could not delete credentials with id '%s'", publicId)));
+                } else {
+                    return Responses.<Boolean>unauthorized("Not allowed to access %s".formatted(publicId));
+                }
+            })
+            .orElseGet(() -> Responses.notFound("%s not found".formatted(publicId)));
     }
 
 }
